feat: restructure API imports and create baseApi for centralized API management
