+++ src/redux/features/jobsApi.tsx
@@ -1,0 +1,13 @@
+import { baseApi } from "./baseApi";
+
+const jobsApi = baseApi.injectEndpoints({
+  endpoints: (builder) => ({
+    createJob: builder.mutation({
+      query: (data) => ({
+        url: "",
+        method: "POST",
+        body: data,
+      }),
+    }),
+  }),
+});
