+++ src/redux/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
 import { rootReducer } from "./rootReducer";
-import { baseApi } from "./api/baseApi";
+import { baseApi } from "./features/baseApi";
 
 const persistConfig = {
   key: "root",
